More refactoring, code cleanup and documentation
